refactor: extract Playable code to separate file
