+++ src/playlist.rs
@@ -4,7 +4,8 @@ use std::sync::Arc;
 use rspotify::model::playlist::{FullPlaylist, SimplifiedPlaylist};
 
 use crate::library::Library;
-use crate::queue::{Playable, Queue};
+use crate::playable::Playable;
+use crate::queue::Queue;
 use crate::spotify::Spotify;
 use crate::track::Track;
 use crate::traits::{IntoBoxedViewExt, ListItem, ViewExt};
