+++ src/ui/listview.rs
@@ -12,7 +12,8 @@ use unicode_width::UnicodeWidthStr;
 use crate::command::{Command, GotoMode, MoveAmount, MoveMode, TargetMode};
 use crate::commands::CommandResult;
 use crate::library::Library;
-use crate::queue::{Playable, Queue};
+use crate::playable::Playable;
+use crate::queue::Queue;
 use crate::track::Track;
 use crate::traits::{IntoBoxedViewExt, ListItem, ViewExt};
 use crate::ui::album::AlbumView;
