+++ src/ui/queue.rs
@@ -9,7 +9,8 @@ use std::sync::Arc;
 use crate::command::{Command, MoveMode, ShiftMode};
 use crate::commands::CommandResult;
 use crate::library::Library;
-use crate::queue::{Playable, Queue};
+use crate::playable::Playable;
+use crate::queue::Queue;
 use crate::traits::ViewExt;
 use crate::ui::listview::ListView;
 use crate::ui::modal::Modal;
